Fixed logic which counts reboot time
Old logic started to count device downtime before execure reboot cmd(it may take some time)
New logic started to count device downtime since SSH connection dropped(device really unreachable)
